requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
itsdangerous==2.2.0
psutil==5.9.8
jinja2==3.1.4
python-multipart==0.0.9
pydantic==1.10.15
Pillow==10.4.0
pillow-heif==0.15.0